Let kotlin tapi x-version tests require target >= 5.3
