--- subprojects/kotlin-dsl-tooling-builders/src/crossVersionTest/groovy/org/gradle/kotlin/dsl/tooling/builders/r47/KotlinSettingsScriptModelCrossVersionSpec.groovy
@@ -1,51 +1,0 @@
-/*
- * Copyright 2019 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.gradle.kotlin.dsl.tooling.builders.r47
-
-import org.gradle.integtests.tooling.fixture.TargetGradleVersion
-import org.gradle.kotlin.dsl.tooling.builders.AbstractKotlinScriptModelCrossVersionTest
-
-import static org.junit.Assert.assertThat
-
-
-@TargetGradleVersion(">=4.7")
-class KotlinSettingsScriptModelCrossVersionSpec extends AbstractKotlinScriptModelCrossVersionTest {
-
-    def "sourcePath includes buildSrc source roots"() {
-
-        given:
-        withKotlinBuildSrc()
-        def settings = withSettings("""include(":sub")""")
-
-        expect:
-        assertThat(
-            sourcePathFor(settings),
-            matchesProjectsSourceRoots(withMainSourceSetJavaKotlinIn("buildSrc")))
-    }
-
-    def "sourcePath includes buildSrc project dependencies source roots"() {
-
-        given:
-        def sourceRoots = withMultiProjectKotlinBuildSrc()
-        def settings = withSettings("""include(":sub")""")
-
-        expect:
-        assertThat(
-            sourcePathFor(settings),
-            matchesProjectsSourceRoots(sourceRoots))
-    }
-}
