+++ subprojects/kotlin-dsl-tooling-builders/src/crossVersionTest/groovy/org/gradle/kotlin/dsl/tooling/builders/r54/KotlinSettingsScriptModelCrossVersionSpec.groovy
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package org.gradle.kotlin.dsl.tooling.builders.r44
+package org.gradle.kotlin.dsl.tooling.builders.r54
 
 import org.gradle.integtests.tooling.fixture.TargetGradleVersion
-
 import org.gradle.kotlin.dsl.tooling.builders.AbstractKotlinScriptModelCrossVersionTest
 
+import static org.junit.Assert.assertThat
 
-@TargetGradleVersion(">=4.4")
+
+@TargetGradleVersion(">=5.3")
 class KotlinSettingsScriptModelCrossVersionSpec extends AbstractKotlinScriptModelCrossVersionTest {
 
     def "can fetch classpath of settings script"() {
@@ -57,5 +58,65 @@ class KotlinSettingsScriptModelCrossVersionSpec extends AbstractKotlinScriptMode
         assertContainsGradleKotlinDslJars(classPath)
         assertIncludes(classPath, settingsDependency)
         assertExcludes(classPath, projectDependency)
+    }
+
+    def "can fetch classpath of settings script plugin"() {
+
+        given:
+        withBuildSrc()
+        withDefaultSettings()
+
+        and:
+        def settingsDependency = withFile("settings-dependency.jar", "")
+        def settings = withFile("my.settings.gradle.kts", """
+            buildscript {
+                dependencies {
+                    classpath(files("${normalizedPathOf(settingsDependency)}"))
+                }
+            }
+        """)
+
+        and:
+        def projectDependency = withFile("project-dependency.jar", "")
+        withFile("build.gradle", """
+            buildscript {
+                dependencies {
+                    classpath(files("${normalizedPathOf(projectDependency)}"))
+                }
+            }
+        """)
+
+        when:
+        def classPath = canonicalClassPathFor(projectDir, settings)
+
+        then:
+        assertContainsBuildSrc(classPath)
+        assertContainsGradleKotlinDslJars(classPath)
+        assertIncludes(classPath, settingsDependency)
+        assertExcludes(classPath, projectDependency)
+    }
+
+    def "sourcePath includes buildSrc source roots"() {
+
+        given:
+        withKotlinBuildSrc()
+        def settings = withSettings("""include(":sub")""")
+
+        expect:
+        assertThat(
+            sourcePathFor(settings),
+            matchesProjectsSourceRoots(withMainSourceSetJavaKotlinIn("buildSrc")))
+    }
+
+    def "sourcePath includes buildSrc project dependencies source roots"() {
+
+        given:
+        def sourceRoots = withMultiProjectKotlinBuildSrc()
+        def settings = withSettings("""include(":sub")""")
+
+        expect:
+        assertThat(
+            sourcePathFor(settings),
+            matchesProjectsSourceRoots(sourceRoots))
     }
 }
